Upgrading Kryo to version 2.24.0
To experiment with running Heritrix3 on platforms like Apache Flink or
Apache Storm, it is necessary to upgrade to a compatible version of
Kryo. This is a significant change as Kryo versions > 1 have changed the
API. In particular, the Kryo instances are no longer thread-safe, which
sigificantly changes how Heritrix3 has to manage things like the
registration of classes for serialization.